Feat: Add restricted token type for users using api token in public clients such as frontend (#1400)
## Objective: 
User's are currently using their workspace token in frontend clients
which is insecure. These workspace tokens allow for creation, listing,
and downloading of artifacts and deployments (user data). We want to
limit the scope of the tokens exposed to the frontend to protect users.

## Current PR:
I implemented a `workspace_restricted` role which limits access to
mostly READ requests on http with some WRITE requests on tasks (for
stopping tasks). This also allows for calling of APIs from the frontend
and prevents calls made to GRPC endpoints.

This is not 100% ideal and I suggest we look into a better RBAC or ACL
implementation so that users can choose of their tokens will work.
However, it is a very simple implementation to quickly get out safer
access to our growing set of users that need access from frontend
clients.

+++ pkg/abstractions/volume/volume.go
@@ -94,6 +94,13 @@ func (vs *GlobalVolumeService) GetOrCreateVolume(ctx context.Context, in *pb.Get
 func (vs *GlobalVolumeService) DeleteVolume(ctx context.Context, in *pb.DeleteVolumeRequest) (*pb.DeleteVolumeResponse, error) {
 	authInfo, _ := auth.AuthInfoFromContext(ctx)
 
+	if !auth.HasPermission(authInfo) {
+		return &pb.DeleteVolumeResponse{
+			Ok:     false,
+			ErrMsg: "Unauthorized Access",
+		}, nil
+	}
+
 	if err := vs.deleteVolume(ctx, authInfo.Workspace, in.Name); err != nil {
 		return &pb.DeleteVolumeResponse{
 			Ok:     false,
